Create BaseTFClient class to consolidate ROS1 and ROS2 impl
