+++ src/tf/TFClient.js
@@ -9,31 +9,16 @@ import Goal from "../actionlib/Goal.js";
 import Service from "../core/Service.js";
 import Topic from "../core/Topic.js";
 
-import Transform from "../math/Transform.js";
-
-import { EventEmitter } from "eventemitter3";
+import BaseTFClient from "./BaseTFClient.js";
 
 /**
  * A TF Client that listens to TFs from tf2_web_republisher.
  */
-export default class TFClient extends EventEmitter {
+export default class TFClient extends BaseTFClient {
   /** @type {Goal|false} */
   currentGoal = false;
   /** @type {Topic|false} */
   currentTopic = false;
-  frameInfos = {};
-  republisherUpdateRequested = false;
-  /** @type {((tf: any) => any) | undefined} */
-  _subscribeCB = undefined;
-  _isDisposed = false;
-  ros;
-  fixedFrame;
-  angularThres;
-  transThres;
-  rate;
-  updateDelay;
-  topicTimeout;
-  serverName;
   repubServiceName;
   actionClient;
   serviceClient;
@@ -51,33 +36,9 @@ export default class TFClient extends EventEmitter {
    * @param {string} [options.serverName="/tf2_web_republisher"] - The name of the tf2_web_republisher server.
    * @param {string} [options.repubServiceName="/republish_tfs"] - The name of the republish_tfs service (non groovy compatibility mode only).
    */
-  constructor({
-    ros,
-    fixedFrame = "base_link",
-    angularThres = 2.0,
-    transThres = 0.01,
-    rate = 10.0,
-    updateDelay = 50,
-    topicTimeout = 2.0,
-    serverName = "/tf2_web_republisher",
-    repubServiceName = "/republish_tfs",
-  }) {
-    super();
+  constructor({ repubServiceName = "/republish_tfs", ...options }) {
+    super(options);
 
-    this.ros = ros;
-    this.fixedFrame = fixedFrame;
-    this.angularThres = angularThres;
-    this.transThres = transThres;
-    this.rate = rate;
-    this.updateDelay = updateDelay;
-    const seconds = topicTimeout;
-    const secs = Math.floor(seconds);
-    const nsecs = Math.floor((seconds - secs) * 1000000000);
-    this.topicTimeout = {
-      secs: secs,
-      nsecs: nsecs,
-    };
-    this.serverName = serverName;
     this.repubServiceName = repubServiceName;
 
     // Create an Action Client
@@ -96,30 +57,7 @@ export default class TFClient extends EventEmitter {
       serviceType: "tf2_web_republisher/RepublishTFs",
     });
   }
-  /**
-   * Process the incoming TF message and send them out using the callback
-   * functions.
-   *
-   * @param {Object} tf - The TF message from the server.
-   */
-  processTFArray(tf) {
-    tf.transforms.forEach((transform) => {
-      let frameID = transform.child_frame_id;
-      if (frameID[0] === "/") {
-        frameID = frameID.substring(1);
-      }
-      const info = this.frameInfos[frameID];
-      if (info) {
-        info.transform = new Transform({
-          translation: transform.transform.translation,
-          rotation: transform.transform.rotation,
-        });
-        info.cbs.forEach((cb) => {
-          cb(info.transform);
-        });
-      }
-    }, this);
-  }
+
   /**
    * Create and send a new goal (or service request) to the tf2_web_republisher
    * based on the current list of TFs.
@@ -163,6 +101,7 @@ export default class TFClient extends EventEmitter {
 
     this.republisherUpdateRequested = false;
   }
+
   /**
    * Process the service response and subscribe to the tf republisher
    * topic.
@@ -195,65 +134,12 @@ export default class TFClient extends EventEmitter {
     // @ts-expect-error Function was bound above
     this.currentTopic.subscribe(this._subscribeCB);
   }
-  /**
-   * @callback subscribeCallback
-   * @param {Transform} callback.transform - The transform data.
-   */
-  /**
-   * Subscribe to the given TF frame.
-   *
-   * @param {string} frameID - The TF frame to subscribe to.
-   * @param {subscribeCallback} callback - Function with the following params:
-   */
-  subscribe(frameID, callback) {
-    // remove leading slash, if it's there
-    if (frameID.startsWith("/")) {
-      frameID = frameID.substring(1);
-    }
-    // if there is no callback registered for the given frame, create empty callback list
-    if (!this.frameInfos[frameID]) {
-      this.frameInfos[frameID] = {
-        cbs: [],
-      };
-      if (!this.republisherUpdateRequested) {
-        setTimeout(this.updateGoal.bind(this), this.updateDelay);
-        this.republisherUpdateRequested = true;
-      }
-    }
 
-    // if we already have a transform, callback immediately
-    else if (this.frameInfos[frameID].transform) {
-      callback(this.frameInfos[frameID].transform);
-    }
-    this.frameInfos[frameID].cbs.push(callback);
-  }
-  /**
-   * Unsubscribe from the given TF frame.
-   *
-   * @param {string} frameID - The TF frame to unsubscribe from.
-   * @param {function} callback - The callback function to remove.
-   */
-  unsubscribe(frameID, callback) {
-    // remove leading slash, if it's there
-    if (frameID.startsWith("/")) {
-      frameID = frameID.substring(1);
-    }
-    const info = this.frameInfos[frameID];
-    // eslint-disable-next-line no-var -- literally what even is going on here
-    for (var cbs = info?.cbs || [], idx = cbs.length; idx--; ) {
-      if (cbs[idx] === callback) {
-        cbs.splice(idx, 1);
-      }
-    }
-    if (!callback || cbs.length === 0) {
-      delete this.frameInfos[frameID];
-    }
-  }
   /**
    * Unsubscribe and unadvertise all topics associated with this TFClient.
    */
   dispose() {
-    this._isDisposed = true;
+    super.dispose();
     this.actionClient.dispose();
     if (this.currentTopic) {
       this.currentTopic.unsubscribe(this._subscribeCB);
